Refactor : Moved to content to CatalogueContent
