+++ app/page.tsx
@@ -1,16 +1,10 @@
 import React from "react"
-import { CatalogueItem, SmoothScroll, CatalogueNavigation, CatalogueFooter } from '@/components'
+import { SmoothScroll, CatalogueFooter, CatalogueContent } from '@/components'
 
-export default function Home() {
-  const itemsArray = Array.from({ length: 60 });
+export default async function Home() {
   return (
     <SmoothScroll>
-      <main className="catalogue-layout">
-        <CatalogueNavigation/>
-        {itemsArray.map((_, index) => (
-          <CatalogueItem key={index} src=""/>
-        ))}
-      </main>
+      <CatalogueContent/>
       <CatalogueFooter/>
     </SmoothScroll>
   );
