+++ components/index.ts
@@ -2,3 +2,4 @@ export { default as CatalogueItem } from './CatalogueItem'
 export { default as SmoothScroll } from './SmoothScroll'
 export { default as CatalogueNavigation } from './CatalogueNavigation'
 export { default as CatalogueFooter } from './CatalogueFooter'
+export { default as CatalogueContent } from './CatalogueContent'
